combining all test cases for HSCS-KEY in one suite
